fix deviation table remove/insert code, reduce tolerance from 1° to 0.3° on saving and add support for external airspeed simulation

+++ main/Deviation.h
@@ -56,7 +56,7 @@ private:
 
 	static xSemaphoreHandle splineMutex;
 	tk::spline *deviationSpline;
-	std::map< double, double> devmap;
+	std::map< int, double> devmap;
 	int _devHolddown;
 	int samples;
 };
